structure of  component  is  break down into  small component and new task  api  is  intigrated
